+++ Imperium/src/Core/Lifecycle/MoonManager.cs
@@ -53,6 +53,7 @@ internal class MoonManager : ImpLifecycleObject
     internal readonly ImpNetworkBinding<bool> TimeIsPaused = new(
         "TimeIsPaused",
         Imperium.Networking,
+        masterBinding: Imperium.Settings.Time.TimeIsPaused,
         onUpdateClient: isPaused =>
         {
             Imperium.TimeOfDay.globalTimeSpeedMultiplier = isPaused ? 0 : ImpConstants.DefaultTimeSpeed;