+++ Imperium/src/Patches/Objects/VehicleControllerPatch.cs
@@ -51,9 +51,6 @@ public static class VehicleControllerPatch
     [HarmonyPatch("Start")]
     internal static void StartPostfixPatch(VehicleController __instance)
     {
-        __instance.mainRigidbody.MovePosition(__instance.transform.position);
-        __instance.hasBeenSpawned = true;
-
         if (Imperium.CruiserManager.SpawnFullTurbo.Value) __instance.AddTurboBoostOnLocalClient(5);
     }